ml/requirements.txt
pandas
scikit-learn
skl2onnx
numpy
psycopg2-binary

